+++ src/eye_extractor/common/drug/antivegf.py
@@ -6,10 +6,10 @@ from eye_extractor.common.drug.shared import build_regex_from_dict, build_patter
 class AntiVegf(enum.IntEnum):
     UNKNOWN = -1
     NONE = 0
-    YES = 1
-    AFLIBERCEPT = 2  # Eyelea
-    BEVACIZUMAB = 3  # Avastin
-    RANIBIZUMAB = 4  # Lucentis
+    YES = 311
+    AFLIBERCEPT = 312  # Eyelea
+    BEVACIZUMAB = 313  # Avastin
+    RANIBIZUMAB = 314  # Lucentis
 
 
 ANTIVEGF_TO_ENUM = {